python: split extension.c into several C files
