+++ python/Makefile
@@ -1,12 +1,16 @@
 
 CFLAGS	= -fPIC -I/usr/include/python2.6 -I../library -Wall -g
-
 PYDIR  ?= /usr/lib64/python2.6/site-packages
+OBJS	= extension.o \
+	  command.o \
+	  status.o \
+	  target.o
+
 
 all:	twopence.so
 
-twopence.so: extension.o
-	$(CC) --shared -o $@ extension.o -L../library -ltwopence
+twopence.so: $(OBJS)
+	$(CC) --shared -o $@ $(OBJS) -L../library -ltwopence
 
 clean:
 	rm -f *.o *.so
