+++ python/extension.c
@@ -1,7 +1,7 @@
 /*
-Twopence python bindings
+Twopence python bindings - extension glue
 
-Copyright (C) 2014 SUSE
+Copyright (C) 2014, 2015 SUSE
 
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -18,177 +18,12 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 */
 
 
-#include <Python.h>
+#include "extension.h"
+
 #include <fcntl.h>
 #include <sys/wait.h>
+
 #include "twopence.h"
-
-typedef struct {
-	PyObject_HEAD
-
-	struct twopence_target *handle;
-} twopence_Target;
-
-typedef struct {
-	PyObject_HEAD
-
-	char *		command;
-	char *		user;
-	char *		stdinPath;
-	int		suppressOutput;
-	PyObject *	stdout;
-	PyObject *	stderr;
-} twopence_Command;
-
-typedef struct {
-	PyObject_HEAD
-
-	int		remoteStatus;
-	PyObject *	stdout;
-	PyObject *	stderr;
-} twopence_Status;
-
-static void		Target_dealloc(twopence_Target *self);
-static PyObject *	Target_new(PyTypeObject *type, PyObject *args, PyObject *kwds);
-static int		Target_init(twopence_Target *self, PyObject *args, PyObject *kwds);
-static PyObject *	Target_run(PyObject *self, PyObject *args, PyObject *kwds);
-static PyObject *	Target_inject(PyObject *self, PyObject *args, PyObject *kwds);
-static PyObject *	Target_extract(PyObject *self, PyObject *args, PyObject *kwds);
-static void		Command_dealloc(twopence_Command *self);
-static PyObject *	Command_new(PyTypeObject *type, PyObject *args, PyObject *kwds);
-static int		Command_init(twopence_Command *self, PyObject *args, PyObject *kwds);
-static PyObject *	Command_getattr(twopence_Command *self, char *name);
-static int		Command_setattr(twopence_Command *self, char *name, PyObject *);
-static int		Command_Check(PyObject *);
-static int		Command_build(twopence_Command *, twopence_command_t *);
-static PyObject *	Command_suppressOutput(twopence_Command *, PyObject *, PyObject *);
-static void		Status_dealloc(twopence_Status *self);
-static PyObject *	Status_new(PyTypeObject *type, PyObject *args, PyObject *kwds);
-static int		Status_init(twopence_Status *self, PyObject *args, PyObject *kwds);
-static PyObject *	Status_getattr(twopence_Status *self, char *name);
-static int		Status_nonzero(twopence_Status *);
-static PyObject *	callType(PyTypeObject *);
-
-static inline void
-drop_string(char **strp)
-{
-	if (*strp)
-		free(*strp);
-	*strp = NULL;
-}
-
-static inline void
-assign_object(PyObject **var, PyObject *obj)
-{
-	if (obj) {
-		Py_INCREF(obj);
-	}
-	if (*var) {
-		Py_DECREF(*var);
-	}
-	*var = obj;
-}
-
-static inline void
-drop_object(PyObject **objp)
-{
-	if (*objp) {
-		Py_DECREF(*objp);
-	}
-	*objp = NULL;
-}
-
-/*
- * Define the python bindings of class "Target"
- *
- * Create objects using
- *   target = twopence.Target("ssh:somehost");
- *
- * Then invoke methods like this:
- *   target.inject("myhostsfile", "/etc/hosts", mode = 0644, user = "root")
- *
- * Note that errors are not indicated through the return value, but through
- * exceptions.
- */
-static PyMethodDef twopence_targetMethods[] = {
-      {	"run", (PyCFunction) Target_run, METH_VARARGS | METH_KEYWORDS,
-	"Run a command on the SUT"
-      },
-      {	"inject", (PyCFunction) Target_inject, METH_VARARGS | METH_KEYWORDS,
-	"Inject a file to the SUT"
-      },
-      {	"extract", (PyCFunction) Target_extract, METH_VARARGS | METH_KEYWORDS,
-	"Extract a file from the SUT"
-      },
-
-      {	NULL }
-};
-
-static PyTypeObject twopence_TargetType = {
-	PyObject_HEAD_INIT(NULL)
-
-	.tp_name	= "twopence.Target",
-	.tp_basicsize	= sizeof(twopence_Target),
-	.tp_flags	= Py_TPFLAGS_DEFAULT,
-	.tp_doc		= "Twopence target",
-
-	.tp_methods	= twopence_targetMethods,
-	.tp_init	= (initproc) Target_init,
-	.tp_new		= Target_new,
-	.tp_dealloc	= (destructor) Target_dealloc,
-};
-
-/*
- * Define the python bindings of class "Command"
- *
- * Create objects using
- *   command = twopence.Command("/bin/ls");
- *
- * Or like this:
- *   out = bytearray()
- *   err = bytearray()
- *   cmd = twopence.Command("/bin/ls", stdout = out, stderr = err, user = "okir")
- *
- * Supported keyword arguments in the constructor:
- *   user
- *	The user to run this command as; default is "root"
- *   stdin
- *	The file to pass to the command's standard input. Right now,
- *	this only accepts a string specifying a path name. File or buffer
- *	objects are not supported yet.
- *   stdout, stderr:
- *	Buffers to write the respective output streams to.
- *	If not specified, output is written to the python interpreter's stdout.
- *	Pass the None object to suppress output.
- *	If you just specify stdout but not stderr, the two output streams
- *	are combined into one and buffered together.
- *
- * To run this command on the SUT, use
- *   target.run(cmd)
- */
-static PyMethodDef twopence_commandMethods[] = {
-      {	"suppressOutput", (PyCFunction) Command_suppressOutput, METH_VARARGS | METH_KEYWORDS,
-	"Do not display command's output to screen"
-      },
-      {	NULL }
-};
-
-static PyTypeObject twopence_CommandType = {
-	PyObject_HEAD_INIT(NULL)
-
-	.tp_name	= "twopence.Command",
-	.tp_basicsize	= sizeof(twopence_Command),
-	.tp_flags	= Py_TPFLAGS_DEFAULT,
-	.tp_doc		= "Twopence command",
-
-	.tp_methods	= twopence_commandMethods,
-	.tp_init	= (initproc) Command_init,
-	.tp_new		= Command_new,
-	.tp_dealloc	= (destructor) Command_dealloc,
-
-	.tp_getattr	= (getattrfunc) Command_getattr,
-	.tp_setattr	= (setattrfunc) Command_setattr,
-};
 
 /*
  * Methods belonging to the module itself.
@@ -205,7 +40,7 @@ static PyMethodDef twopence_methods[] = {
 /*
  * Convert twopence error to an exception
  */
-static PyObject *
+PyObject *
 twopence_Exception(const char *msg, int rc)
 {
 	char buffer[256];
@@ -215,645 +50,9 @@ twopence_Exception(const char *msg, int rc)
 	return NULL;
 }
 
-/*
- * Constructor: allocate empty Command object, and set its members.
- */
-static PyObject *
-Command_new(PyTypeObject *type, PyObject *args, PyObject *kwds)
-{
-	twopence_Command *self;
-
-	self = (twopence_Command *) type->tp_alloc(type, 0);
-	if (self == NULL)
-		return NULL;
-
-	/* init members */
-	self->command = NULL;
-	self->user = NULL;
-	self->stdinPath = NULL;
-	self->stdout = NULL;
-	self->stderr = NULL;
-	self->suppressOutput = 0;
-
-	return (PyObject *)self;
-}
-
-/*
- * Initialize the command object
- *
- * Typical ways to do this include
- *    cmd = twopence.Command("/bin/ls", user = "okir", stdout = bytearray());
- *    cmd = twopence.Command("/bin/ls", user = "okir", stdout = str());
- *    cmd = twopence.Command("/usr/bin/wc", stdin = "/etc/hosts");
- */
-static int
-Command_init(twopence_Command *self, PyObject *args, PyObject *kwds)
-{
-	static char *kwlist[] = {
-		"command",
-		"user",
-		"stdin",
-		"stdout",
-		"stderr",
-		NULL
-	};
-	PyObject *stdinObject = NULL, *stdoutObject = NULL, *stderrObject = NULL;
-	char *command, *user = NULL;
-
-	if (!PyArg_ParseTupleAndKeywords(args, kwds, "s|sOOO", kwlist, &command, &user, &stdinObject, &stdoutObject, &stderrObject))
-		return -1;
-
-	self->command = strdup(command);
-	self->user = user? strdup(user) : NULL;
-	self->stdout = NULL;
-	self->stderr = NULL;
-	self->stdinPath = NULL;
-	self->suppressOutput = 0;
-
-	if (stdoutObject == NULL) {
-		stdoutObject = callType(&PyByteArray_Type);
-	} else {
-		Py_INCREF(stdoutObject);
-	}
-	self->stdout = stdoutObject;
-
-	if (stderrObject == NULL)
-		stderrObject = stdoutObject;
-
-	Py_INCREF(stderrObject);
-	self->stderr = stderrObject;
-
-	if (stdinObject && stdinObject != Py_None) {
-		char *s;
-
-		if ((s = PyString_AsString(stdinObject)) == NULL)
-			return -1;
-		self->stdinPath = strdup(s);
-	}
-
-	return 0;
-}
-
-/*
- * Destructor: clean any state inside the Command object
- */
 static void
-Command_dealloc(twopence_Command *self)
+twopence_registerType(PyObject *m, const char *name, PyTypeObject *type)
 {
-	drop_string(&self->command);
-	drop_string(&self->user);
-	drop_string(&self->stdinPath);
-	drop_object(&self->stdout);
-	drop_object(&self->stderr);
-}
-
-static int
-Command_Check(PyObject *self)
-{
-	return PyType_IsSubtype(Py_TYPE(self), &twopence_CommandType);
-}
-
-static int
-Command_build(twopence_Command *self, twopence_command_t *cmd)
-{
-	twopence_buffer_t *buffer = NULL;
-
-	twopence_command_init(cmd, self->command);
-
-	cmd->user = self->user;
-
-	twopence_command_ostreams_reset(cmd);
-	if (self->suppressOutput || self->stdout == Py_None) {
-		/* ostream has already been reset above */
-	} else {
-		/* Copy remote stdout to our stdout */
-		twopence_command_iostream_redirect(cmd, TWOPENCE_STDOUT, 1, false);
-	}
-
-	if (self->stdout == NULL || PyByteArray_Check(self->stdout)) {
-		/* Capture command output in a buffer */
-		buffer = twopence_command_alloc_buffer(cmd, TWOPENCE_STDOUT, 65536);
-		twopence_command_ostream_capture(cmd, TWOPENCE_STDOUT, buffer);
-	}
-
-	if (self->suppressOutput || self->stderr == Py_None) {
-		/* ostream has already been reset above */
-	} else {
-		/* Copy remote stderr to our stderr */
-		twopence_command_iostream_redirect(cmd, TWOPENCE_STDERR, 2, false);
-	}
-
-	/* If cmd.stdout and cmd.stderr are both NULL, or both refer to the same
-	 * bytearray object, send the remote stdout and stderr to a shared buffer */
-	if (buffer != NULL && self->stderr == self->stdout) {
-		twopence_command_ostream_capture(cmd, TWOPENCE_STDERR, buffer);
-	} else 
-	if (self->stderr == NULL || PyByteArray_Check(self->stderr)) {
-		buffer = twopence_command_alloc_buffer(cmd, TWOPENCE_STDERR, 65536);
-		twopence_command_ostream_capture(cmd, TWOPENCE_STDERR, buffer);
-	}
-
-	if (self->stdinPath != NULL) {
-		int fd = open(self->stdinPath, O_RDONLY);
-
-		twopence_command_iostream_redirect(cmd, TWOPENCE_STDIN, fd, true);
-	}
-
-	return 0;
-}
-
-static PyObject *
-Command_suppressOutput(twopence_Command *self, PyObject *args, PyObject *kwds)
-{
-	static char *kwlist[] = {
-		NULL
-	};
-	if (!PyArg_ParseTupleAndKeywords(args, kwds, "", kwlist))
-		return NULL;
-
-	self->suppressOutput = 1;
-	Py_INCREF(Py_None);
-	return Py_None;
-}
-
-static PyObject *
-Command_stdout(twopence_Command *self)
-{
-	PyObject *result;
-
-	result = self->stdout;
-	if (result == NULL)
-		result = Py_None;
-	Py_INCREF(result);
-	return result;
-}
-
-static PyObject *
-Command_stderr(twopence_Command *self)
-{
-	PyObject *result;
-
-	result = self->stderr;
-	if (result == NULL)
-		result = Py_None;
-	Py_INCREF(result);
-	return result;
-}
-
-static PyObject *
-Command_getattr(twopence_Command *self, char *name)
-{
-	if (!strcmp(name, "stdout"))
-		return Command_stdout(self);
-	if (!strcmp(name, "stderr"))
-		return Command_stderr(self);
-
-	return Py_FindMethod(twopence_commandMethods, (PyObject *) self, name);
-}
-
-static int
-Command_setattr(twopence_Command *self, char *name, PyObject *v)
-{
-	if (!strcmp(name, "stdout")) {
-		if (v != Py_None && !PyByteArray_Check(v))
-			goto bad_attr;
-		assign_object(&self->stdout, v);
-		return 0;
-	}
-	if (!strcmp(name, "stderr")) {
-		if (v != Py_None && !PyByteArray_Check(v))
-			goto bad_attr;
-		assign_object(&self->stderr, v);
-		return 0;
-	}
-
-	(void) PyErr_Format(PyExc_AttributeError, "Unknown attribute: %s", name);
-	return -1;
-
-bad_attr:
-	(void) PyErr_Format(PyExc_AttributeError, "Incompatible value for attribute: %s", name);
-	return -1;
-
-}
-
-/*
- * Command status
- */
-/*
- * Define the python bindings of class "Status"
- * Normally, you do not create Status objects yourself;
- * Usually, these are created as the return value of Command.run()
- */
-static PyMethodDef twopence_statusMethods[] = {
-      {	NULL }
-};
-
-static PyNumberMethods twopence_statusAsNumber = {
-	.nb_nonzero	= (inquiry) Status_nonzero,
-};
-
-static PyTypeObject twopence_StatusType = {
-	PyObject_HEAD_INIT(NULL)
-
-	.tp_name	= "twopence.Status",
-	.tp_basicsize	= sizeof(twopence_Status),
-	.tp_flags	= Py_TPFLAGS_DEFAULT,
-	.tp_doc		= "Twopence status",
-
-	.tp_methods	= twopence_statusMethods,
-	.tp_init	= (initproc) Status_init,
-	.tp_new		= Status_new,
-	.tp_dealloc	= (destructor) Status_dealloc,
-
-	.tp_getattr	= (getattrfunc) Status_getattr,
-	.tp_as_number	= &twopence_statusAsNumber,
-};
-
-/*
- * Constructor: allocate empty Status object, and set its members.
- */
-static PyObject *
-Status_new(PyTypeObject *type, PyObject *args, PyObject *kwds)
-{
-	twopence_Status *self;
-
-	self = (twopence_Status *) type->tp_alloc(type, 0);
-	if (self == NULL)
-		return NULL;
-
-	/* init members */
-	self->remoteStatus = 0;
-	self->stdout = NULL;
-	self->stderr = NULL;
-
-	return (PyObject *)self;
-}
-
-/*
- * Initialize the status object
- */
-static int
-Status_init(twopence_Status *self, PyObject *args, PyObject *kwds)
-{
-	static char *kwlist[] = {
-		"status",
-		"stdout",
-		"stderr",
-		NULL
-	};
-	PyObject *stdoutObject = NULL, *stderrObject = NULL;
-	int exitval = 0;
-
-	if (args == Py_None)
-		return 0;
-
-	if (!PyArg_ParseTupleAndKeywords(args, kwds, "|iOO", kwlist, &exitval, &stdoutObject, &stderrObject))
-		return -1;
-
-	self->remoteStatus = exitval;
-	self->stdout = NULL;
-	self->stderr = NULL;
-
-	if (stdoutObject) {
-		Py_INCREF(stdoutObject);
-		self->stdout = stdoutObject;
-	}
-	if (stderrObject) {
-		Py_INCREF(stderrObject);
-		self->stderr = stderrObject;
-	}
-
-	return 0;
-}
-
-/*
- * Destructor: clean any state inside the Status object
- */
-static void
-Status_dealloc(twopence_Status *self)
-{
-	drop_object(&self->stdout);
-	drop_object(&self->stderr);
-}
-
-int
-Status_Check(PyObject *self)
-{
-	return PyType_IsSubtype(Py_TYPE(self), &twopence_StatusType);
-}
-
-static PyObject *
-Status_stdout(twopence_Status *self)
-{
-	PyObject *result;
-
-	result = self->stdout;
-	if (result == NULL)
-		result = Py_None;
-	Py_INCREF(result);
-	return result;
-}
-
-static PyObject *
-Status_stderr(twopence_Status *self)
-{
-	PyObject *result;
-
-	result = self->stderr;
-	if (result == NULL)
-		result = Py_None;
-	Py_INCREF(result);
-	return result;
-}
-
-static PyObject *
-Status_message(twopence_Status *self)
-{
-	char message[128];
-
-	message[0] = '\0';
-
-	/* Unfortunately, the exit status returned by libssh is somewhat limited :-( */
-	switch (self->remoteStatus) {
-	case 0:
-		strcpy(message, "success");
-		break;
-
-	case -1:
-		strcpy(message, "crashed");
-		break;
-
-	default:
-		snprintf(message, sizeof(message), "status %d", self->remoteStatus);
-	}
-
-	return PyString_FromString(message);
-}
-
-static PyObject *
-Status_getattr(twopence_Status *self, char *name)
-{
-	if (!strcmp(name, "stdout"))
-		return Status_stdout(self);
-	if (!strcmp(name, "stderr"))
-		return Status_stderr(self);
-	if (!strcmp(name, "code"))
-		return PyInt_FromLong(self->remoteStatus);
-	if (!strcmp(name, "message"))
-		return Status_message(self);
-
-	PyErr_Format(PyExc_AttributeError, "%s", name);
-	return NULL;
-}
-
-/*
- * Support using a status object in a boolean context, as in
- *
- * status = target.run(cmd)
- * if status:
- *    all is well
- *
- */
-static int
-Status_nonzero(twopence_Status *self)
-{
-	return self->remoteStatus == 0;
-}
-
-/*
- * Constructor: allocate empty Target object, and set its members.
- */
-static PyObject *
-Target_new(PyTypeObject *type, PyObject *args, PyObject *kwds)
-{
-	twopence_Target *self;
-
-	self = (twopence_Target *) type->tp_alloc(type, 0);
-	if (self == NULL)
-		return NULL;
-
-	/* init members */
-	self->handle = NULL;
-
-	return (PyObject *)self;
-}
-
-static int
-Target_init(twopence_Target *self, PyObject *args, PyObject *kwds)
-{
-	char *targetSpec;
-	int rc;
-
-	static char *kwlist[] = {"target", NULL};
-
-	if (!PyArg_ParseTupleAndKeywords(args, kwds, "s", kwlist, &targetSpec))
-		return -1; 
-
-	rc = twopence_target_new(targetSpec, &self->handle);
-	if (rc < 0) {
-		twopence_Exception("Target initialization", rc);
-		return -1;
-	}
-
-	return 0;
-}
-
-/*
- * Destructor: clean any state inside the Target object
- */
-static void
-Target_dealloc(twopence_Target *self)
-{
-	if (self->handle)
-		twopence_target_free(self->handle);
-	self->handle = NULL;
-}
-
-/*
- * Extract twopence target handle from python object.
- * This should really do a type check and throw an exception if it doesn't match
- */
-static struct twopence_target *
-Target_handle(PyObject *self)
-{
-	return ((twopence_Target *) self)->handle;
-}
-
-/*
- * The run() method can return data in buffers
- * Support this here.
- */
-int
-twopence_AppendBuffer(PyObject *buffer, const twopence_buffer_t *buf)
-{
-	int rv = 0;
-
-	if (buffer != NULL && buffer != Py_None && buf->head != NULL) {
-		PyObject *temp = PyString_FromString(buf->head);
-
-		if (PySequence_InPlaceConcat(buffer, temp) == NULL)
-			rv = -1;
-		Py_DECREF(temp);
-	}
-	return rv;
-}
-
-/*
- * Run a command on the SUT
- *
- * By default, this will print stderr and stdout to the screen (ie the client
- * side's stdout).
- *
- * If you want to capture all output for later processing, you can pass one
- * or two string objects as parameters 'stdout' und 'stderr'.
- *
- * To suppress all output, pass "none" objects to 'stdout' und 'stderr'.
- */
-static PyObject *
-Target_run(PyObject *self, PyObject *args, PyObject *kwds)
-{
-	struct twopence_target *handle;
-	twopence_Command *cmdObject = NULL;
-	twopence_Status *statusObject;
-	twopence_command_t cmd;
-	twopence_status_t status;
-	PyObject *result = NULL;
-	int rc;
-
-	memset(&cmd, 0, sizeof(cmd));
-
-	if (PySequence_Check(args)
-	 && PySequence_Fast_GET_SIZE(args) == 1) {
-		/* Single argument can be an object of type Command or a string */
-		PyObject *object = PySequence_Fast_GET_ITEM(args, 0);
-
-		if (Command_Check(object)) {
-			cmdObject = (twopence_Command *) object;
-			Py_INCREF(cmdObject);
-		}
-	}
-
-	if (cmdObject == NULL) {
-		cmdObject = PyObject_New(twopence_Command, &twopence_CommandType);
-
-		if (Command_init(cmdObject, args, kwds) < 0)
-			goto out;
-	}
-
-	if ((handle = Target_handle(self)) == NULL)
-		goto out;
-
-	if (Command_build(cmdObject, &cmd) < 0)
-		goto out;
-
-	printf("run \"%s\" as user %s\n", cmd.command, cmd.user);
-	rc = twopence_run_test(handle, &cmd, &status);
-	if (rc < 0) {
-		twopence_Exception("run", rc);
-		goto out;
-	}
-
-	/* Now funnel the captured data to the respective buffer objects */
-	if (twopence_AppendBuffer(cmdObject->stdout, &cmd.buffer[TWOPENCE_STDOUT]) < 0)
-		goto out;
-	if (twopence_AppendBuffer(cmdObject->stderr, &cmd.buffer[TWOPENCE_STDERR]) < 0)
-		goto out;
-
-	statusObject = (twopence_Status *) callType(&twopence_StatusType);
-	statusObject->remoteStatus = status.minor;
-	if (cmdObject->stdout) {
-		statusObject->stdout = cmdObject->stdout;
-		Py_INCREF(statusObject->stdout);
-	}
-	if (cmdObject->stderr) {
-		statusObject->stderr = cmdObject->stderr;
-		Py_INCREF(statusObject->stderr);
-	}
-
-	result = (PyObject *) statusObject;
-
-out:
-	if (cmdObject) {
-		Py_DECREF(cmdObject);
-	}
-
-	twopence_command_destroy(&cmd);
-	return result;
-}
-
-/*
- * inject file into SUT
- */
-static PyObject *
-Target_inject(PyObject *self, PyObject *args, PyObject *kwds)
-{
-	static char *kwlist[] = {
-		"local",
-		"remote",
-		"user",
-		"mode",
-		NULL
-	};
-	struct twopence_target *handle;
-	char *sourceFile, *destFile;
-	char *user = "root";
-	int omode = 0644;
-	int rc, remoteRc;
-
-	if (!PyArg_ParseTupleAndKeywords(args, kwds, "ss|si", kwlist, &sourceFile, &destFile, &user, &omode))
-		return NULL;
-
-	printf("inject %s -> %s (user %s, mode 0%o)\n", sourceFile, destFile, user, omode);
-
-	if ((handle = Target_handle(self)) == NULL)
-		return NULL;
-
-	rc = twopence_inject_file(handle, user, sourceFile, destFile, &remoteRc, 0);
-	if (rc < 0)
-		return twopence_Exception("inject", rc);
-
-	Py_INCREF(Py_None);
-	return Py_None;
-}
-
-/*
- * extract file from SUT
- */
-static PyObject *
-Target_extract(PyObject *self, PyObject *args, PyObject *kwds)
-{
-	static char *kwlist[] = {
-		"local",
-		"remote",
-		"user",
-		"mode",
-		NULL
-	};
-	struct twopence_target *handle;
-	char *sourceFile, *destFile;
-	char *user = "root";
-	int omode = 0644;
-	int rc, remoteRc;
-
-	if (!PyArg_ParseTupleAndKeywords(args, kwds, "ss|si", kwlist, &sourceFile, &destFile, &user, &omode))
-		return NULL;
-
-	printf("extract %s -> %s (user %s, mode 0%o)\n", sourceFile, destFile, user, omode);
-	if ((handle = Target_handle(self)) == NULL)
-		return NULL;
-
-	rc = twopence_extract_file(handle, user, sourceFile, destFile, &remoteRc, 0);
-	if (rc < 0)
-		return twopence_Exception("extract", rc);
-
-	Py_INCREF(Py_None);
-	return Py_None;
-}
-
-static void
-registerType(PyObject *m, const char *name, PyTypeObject *type)
-{
-	//type->tp_new = PyType_GenericNew;
 	if (PyType_Ready(type) < 0)
 		return;
 
@@ -861,14 +60,18 @@ registerType(PyObject *m, const char *name, PyTypeObject *type)
 	PyModule_AddObject(m, name, (PyObject *) type);
 }
 
-static PyObject *
-callType(PyTypeObject *typeObject)
+PyObject *
+twopence_callType(PyTypeObject *typeObject, PyObject *args, PyObject *kwds)
 {
-	PyObject *args = PyTuple_New(0);
 	PyObject *obj;
 
-	obj = PyObject_Call((PyObject *) typeObject, args, NULL);
-	Py_DECREF(args);
+	if (args == NULL) {
+		args = PyTuple_New(0);
+		obj = PyObject_Call((PyObject *) typeObject, args, NULL);
+		Py_DECREF(args);
+	} else {
+		obj = PyObject_Call((PyObject *) typeObject, args, kwds);
+	}
 
 	return obj;
 }
@@ -880,7 +83,7 @@ inittwopence(void)
 
 	m = Py_InitModule3("twopence", twopence_methods, "Module for twopence based testing");
 
-	registerType(m, "Target", &twopence_TargetType);
-	registerType(m, "Command", &twopence_CommandType);
-	registerType(m, "Status", &twopence_StatusType);
+	twopence_registerType(m, "Target", &twopence_TargetType);
+	twopence_registerType(m, "Command", &twopence_CommandType);
+	twopence_registerType(m, "Status", &twopence_StatusType);
 }
